I think I might have fixed the issue
In the last commit before I reverted thecommit, I noticed that Git said that there were 16 additions and 16 deletions. I thought that was odd, so I looked at the 32-bit commit and that one had only 2 insertions and 2 deletions. I hope that was an accident and not Visual Studio's fault.